fix(train): gate aux losses + AdamW for ablations + bf16-safe rfft
Root cause of stalled lm_loss: Adafactor warmup_init=True keeps the
effective LR near 1e-4 with extra per-param RMS scaling, so a 1000-step
ablation never reaches a learning LR. Compounded by every auxiliary
loss (world/motor/phi/cpc/kl/novel/id_drift/pred_coding) running at
full cfg weight from step 0 against a random-init network, plus the
homeostasis controller correcting NT levels toward targets calibrated
for a trained network -- driving 5HT/GABA to ceiling.

Changes:
- brain.Brain: new `_aux_w_scale` and `_infancy` flags. forward_lm now
  multiplies every non-LM loss term by _aux_w_scale, and skips
  trophic.update + language.bdnf_grow_all while _infancy is True.
- train.py: per-step `brain._aux_w_scale = _aux_w` / `brain._infancy
  = in_infancy`; removes the old single-aux mutation of
  cfg.w_pred_coding (replaced by the unified scale).
- train.py: homeostasis.observe gated on `_maturation_awakened` so it
  doesn't run during random-init phase.
- train.py: running grad-norm accumulated and shown in the log line
  (both baseline and full).
- config.BrainConfig: w_lm 2.0 -> 1.0 (the doubling inflated displayed
  total without speeding LM convergence).
- intelligence/oscillations.py: rfft upcast input to float32 (bfloat16
  not supported by torch.fft.rfft).
- colab_run.ipynb cell 4: GRAD_ACCUM 4 -> 16 (effective batch 32K
  tokens/step -- healthier signal for 228M params).
- colab_run.ipynb cell 5: ablation commands pass --optimizer adamw so
  cfg.lr=3e-4 + 200-step warmup + cosine actually takes effect.

+++ neuroslm/intelligence/oscillations.py
@@ -180,8 +180,8 @@ class NeuralOscillationTracker(nn.Module):
             if effective_w < 4:
                 continue
 
-            # FFT
-            spectrum = torch.fft.rfft(rms)
+            # FFT — torch.fft.rfft does not support bfloat16, upcast to float32
+            spectrum = torch.fft.rfft(rms.float())
             power = spectrum.abs().pow(2)
 
             # Band extraction (map to FFT bins)
